+++ data/src/main/java/gr/eduinvoice/data/user/CurrentUserProvider.kt
@@ -1,0 +1,7 @@
+package gr.eduinvoice.data.user
+
+import kotlinx.coroutines.flow.Flow
+
+interface CurrentUserProvider {
+    val loggedInUserId: Flow<Long?>
+}